Refactor into smaller classes
MOL-95 MOL-146
