+++ resources/js/applePayRequest.js
@@ -1,0 +1,23 @@
+export const request = (countryCode, currencyCode, totalLabel, subtotal) => {
+    return {
+        countryCode: countryCode,
+        currencyCode: currencyCode,
+        supportedNetworks: ['amex', 'maestro', 'masterCard', 'visa', 'vPay'],
+        merchantCapabilities: ['supports3DS'],
+        shippingType: 'shipping',
+        requiredBillingContactFields: [
+            'postalAddress',
+            'email'
+        ],
+        requiredShippingContactFields: [
+            'postalAddress',
+            'email'
+        ],
+        total: {
+            label: totalLabel,
+            amount: subtotal,
+            type: 'pending'
+        }
+    }
+}
+
