+++ src/Mollie/WC/ApplePayButton/DataToAppleButtonScripts.php
@@ -1,0 +1,128 @@
+<?php
+
+class Mollie_WC_ApplePayButton_DataToAppleButtonScripts
+{
+    /**
+     * Sets the appropriate data to send to ApplePay script
+     * Data differs between product page and cart page
+     *
+     * @return array|bool
+     */
+    public function applePayScriptData()
+    {
+        $base_location = wc_get_base_location();
+        $shopCountryCode = $base_location['country'];
+        $currencyCode = get_woocommerce_currency();
+        $totalLabel = get_bloginfo('name');
+        if (is_product()) {
+            return $this->dataForProductPage(
+                $shopCountryCode,
+                $currencyCode,
+                $totalLabel
+            );
+        }
+        if (is_cart()) {
+            return $this->dataForCartPage(
+                $shopCountryCode,
+                $currencyCode,
+                $totalLabel
+            );
+        }
+        return [];
+    }
+
+    /**
+     * Check if the product needs shipping
+     *
+     * @param $product
+     *
+     * @return bool
+     */
+    protected function checkIfNeedShipping($product)
+    {
+        if (!wc_shipping_enabled()
+            || 0 === wc_get_shipping_method_count(
+                true
+            )
+        ) {
+            return false;
+        }
+        $needs_shipping = false;
+
+        if ($product->needs_shipping()) {
+            $needs_shipping = true;
+        }
+
+        return $needs_shipping;
+    }
+
+    /**
+     * @param $shopCountryCode
+     * @param $currencyCode
+     * @param $totalLabel
+     *
+     * @return array|bool
+     */
+    protected function dataForProductPage(
+        $shopCountryCode,
+        $currencyCode,
+        $totalLabel
+    ) {
+        $product = wc_get_product(get_the_id());
+        if (!$product) {
+            return false;
+        }
+        $isVariation = false;
+        if ($product->get_type() === 'variable') {
+            $isVariation = true;
+        }
+        $productNeedShipping = $this->checkIfNeedShipping($product);
+        $productId = get_the_id();
+        $productPrice = $product->get_price();
+
+        return [
+            'product' => [
+                'needShipping' => $productNeedShipping,
+                'id' => $productId,
+                'price' => $productPrice,
+                'isVariation' => $isVariation,
+            ],
+            'shop' => [
+                'countryCode' => $shopCountryCode,
+                'currencyCode' => $currencyCode,
+                'totalLabel' => $totalLabel
+            ],
+            'ajaxUrl' => admin_url('admin-ajax.php')
+        ];
+    }
+
+    /**
+     * @param $shopCountryCode
+     * @param $currencyCode
+     * @param $totalLabel
+     *
+     * @return array
+     */
+    protected function dataForCartPage(
+        $shopCountryCode,
+        $currencyCode,
+        $totalLabel
+    ) {
+        $cart = WC()->cart;
+
+        return [
+            'product' => [
+                'needShipping' => $cart->needs_shipping(),
+                'subtotal' => $cart->get_subtotal(),
+
+            ],
+            'shop' => [
+                'countryCode' => $shopCountryCode,
+                'currencyCode' => $currencyCode,
+                'totalLabel' => $totalLabel
+            ],
+            'ajaxUrl' => admin_url('admin-ajax.php')
+        ];
+    }
+
+}
